refactor:remove victord
