--- daemon/.env
@@ -1,2 +1,0 @@
-HOST=127.0.0.1
-PORT=8080
