+++ go.mod
@@ -1,9 +1,3 @@
 module victorgo
 
 go 1.24.1
-
-require (
-	github.com/google/uuid v1.6.0
-	github.com/gorilla/mux v1.8.1
-	github.com/joho/godotenv v1.4.0
-)
